seems I have got rid of some bugs and its passing

+++ test/Examples/get_apr.cpp
@@ -157,7 +157,7 @@ int main(int argc, char **argv) {
     for(int depth = part_map.k_min; depth <= (part_map.k_max+1);depth++){
         
         for(uint64_t i = 0; i < part_map.downsampled[depth].mesh.size();i++){
-            part_map.downsampled[depth].mesh[i] = i;
+            part_map.downsampled[depth].mesh[i] = (uint16_t) i;
         }
         
     }
